macOS: fold OSSurfaceView into SurfaceView
OSSurfaceView existed as a platform-neutral base class for the iOS port,
which never landed. Everything it declared was either overridden by
SurfaceView or only meaningful with a live `ghostty_surface_t`, so the
base class was an indirection with exactly one subclass:

- `surface` returned nil and was overridden to read `surfaceModel`.
- `focusDidChange`/`sizeDidChange` were empty placeholders, overridden.
- `endSearch` was overridden to move focus before calling `super`.
- `progressReport`/`searchState` were plain `@Published` properties that
  SurfaceView overrode only to attach `didSet` observers.

Move the state, the readonly notification observer, `highlight()`,
`setChildExitedMessage`, and the nested `SearchState` type into
SurfaceView and drop the `override`s. `id` is now assigned directly in
SurfaceView's init instead of being threaded through `super.init(id:)`.

No behavior change: the readonly observer is registered alongside the
other notification observers in the same init, and SurfaceView's `deinit`
already removed all observers.
